+++ beacon-chain/sync/subscriber_beacon_blocks_test.go
@@ -9,14 +9,20 @@ import (
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/blockchain"
 	chainMock "github.com/prysmaticlabs/prysm/v5/beacon-chain/blockchain/testing"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/core/helpers"
+	"github.com/prysmaticlabs/prysm/v5/beacon-chain/db/filesystem"
 	dbtest "github.com/prysmaticlabs/prysm/v5/beacon-chain/db/testing"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/execution"
+	mockExecution "github.com/prysmaticlabs/prysm/v5/beacon-chain/execution/testing"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/operations/attestations"
+	mockp2p "github.com/prysmaticlabs/prysm/v5/beacon-chain/p2p/testing"
+	"github.com/prysmaticlabs/prysm/v5/beacon-chain/startup"
 	lruwrpr "github.com/prysmaticlabs/prysm/v5/cache/lru"
+	"github.com/prysmaticlabs/prysm/v5/consensus-types/blocks"
 	ethpb "github.com/prysmaticlabs/prysm/v5/proto/prysm/v1alpha1"
 	"github.com/prysmaticlabs/prysm/v5/testing/assert"
 	"github.com/prysmaticlabs/prysm/v5/testing/require"
 	"github.com/prysmaticlabs/prysm/v5/testing/util"
+	"github.com/prysmaticlabs/prysm/v5/time"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -66,7 +72,9 @@ func TestService_beaconBlockSubscriber(t *testing.T) {
 						DB:   db,
 						Root: make([]byte, 32),
 					},
-					attPool: attestations.NewPool(),
+					attPool:                attestations.NewPool(),
+					blobStorage:            filesystem.NewEphemeralBlobStorage(t),
+					executionReconstructor: &mockExecution.EngineClient{},
 				},
 			}
 			s.initCaches()
@@ -123,4 +131,66 @@ func TestService_BeaconBlockSubscribe_UndefinedEeError(t *testing.T) {
 	require.ErrorIs(t, s.beaconBlockSubscriber(context.Background(), util.NewBeaconBlock()), blockchain.ErrUndefinedExecutionEngineError)
 	require.Equal(t, 0, len(s.badBlockCache.Keys()))
 	require.Equal(t, 1, len(s.seenBlockCache.Keys()))
+}
+
+func TestReconstructAndBroadcastBlobs(t *testing.T) {
+	rob, err := blocks.NewROBlob(
+		&ethpb.BlobSidecar{
+			SignedBlockHeader: &ethpb.SignedBeaconBlockHeader{
+				Header: &ethpb.BeaconBlockHeader{
+					ParentRoot: make([]byte, 32),
+					BodyRoot:   make([]byte, 32),
+					StateRoot:  make([]byte, 32),
+				},
+				Signature: []byte("signature"),
+			},
+		})
+	require.NoError(t, err)
+
+	chainService := &chainMock.ChainService{
+		Genesis: time.Now(),
+	}
+
+	b := util.NewBeaconBlockDeneb()
+	sb, err := blocks.NewSignedBeaconBlock(b)
+	require.NoError(t, err)
+
+	tests := []struct {
+		name              string
+		blobSidecars      []blocks.VerifiedROBlob
+		expectedBlobCount int
+	}{
+		{
+			name:              "Constructed 0 blobs",
+			blobSidecars:      nil,
+			expectedBlobCount: 0,
+		},
+		{
+			name: "Constructed 6 blobs",
+			blobSidecars: []blocks.VerifiedROBlob{
+				{ROBlob: rob}, {ROBlob: rob}, {ROBlob: rob}, {ROBlob: rob}, {ROBlob: rob}, {ROBlob: rob},
+			},
+			expectedBlobCount: 6,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			s := Service{
+				cfg: &config{
+					p2p:         mockp2p.NewTestP2P(t),
+					chain:       chainService,
+					clock:       startup.NewClock(time.Now(), [32]byte{}),
+					blobStorage: filesystem.NewEphemeralBlobStorage(t),
+					executionReconstructor: &mockExecution.EngineClient{
+						BlobSidecars: tt.blobSidecars,
+					},
+					operationNotifier: &chainMock.MockOperationNotifier{},
+				},
+				seenBlobCache: lruwrpr.New(1),
+			}
+			s.reconstructAndBroadcastBlobs(context.Background(), sb)
+			require.Equal(t, tt.expectedBlobCount, len(chainService.Blobs))
+		})
+	}
 }